refactor: LineResponses dto 추가 및 dto 패키지 구분
