+++ src/main/java/wooteco/subway/line/LineController.java
@@ -2,10 +2,11 @@ package wooteco.subway.line;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import wooteco.subway.line.dto.LineRequest;
+import wooteco.subway.line.dto.LineResponse;
+import wooteco.subway.line.dto.LineResponses;
 
 import java.net.URI;
-import java.util.List;
-import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/lines")
@@ -25,32 +26,27 @@ public class LineController {
     }
 
     @GetMapping
-    public ResponseEntity<List<LineResponse>> showLines() {
-        List<Line> lines = lineService.findAll();
-        List<LineResponse> lineResponses = lines.stream()
-                                                .map(line -> new LineResponse(line.getId(), line.getName(), line.getColor()))
-                                                .collect(Collectors.toList());
-        return ResponseEntity.ok()
-                             .body(lineResponses);
+    public ResponseEntity<LineResponses> showLines() {
+        LineResponses lineResponses = LineResponses.of(lineService.findAll());
+        return ResponseEntity.ok(lineResponses);
     }
 
     @GetMapping("/{id}")
     public ResponseEntity<LineResponse> showLine(@PathVariable Long id) {
         Line findLine = lineService.findById(id);
         LineResponse lineResponse = new LineResponse(findLine.getId(), findLine.getName(), findLine.getColor());
-        return ResponseEntity.ok()
-                             .body(lineResponse);
+        return ResponseEntity.ok(lineResponse);
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity modifyLine(@PathVariable Long id, @RequestBody LineRequest lineRequest) {
+    public ResponseEntity<Void> modifyLine(@PathVariable Long id, @RequestBody LineRequest lineRequest) {
         lineService.modifyLine(id, lineRequest.getName(), lineRequest.getColor());
         return ResponseEntity.ok()
                              .build();
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity deleteLine(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteLine(@PathVariable Long id) {
         lineService.deleteLine(id);
         return ResponseEntity.noContent()
                              .build();
