+++ src/main/java/wooteco/subway/station/StationController.java
@@ -3,6 +3,8 @@ package wooteco.subway.station;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import wooteco.subway.station.dto.StationRequest;
+import wooteco.subway.station.dto.StationResponse;
 
 import java.net.URI;
 import java.util.List;
@@ -30,12 +32,11 @@ public class StationController {
         List<StationResponse> stationResponses = stations.stream()
                                                          .map(it -> new StationResponse(it.getId(), it.getName()))
                                                          .collect(Collectors.toList());
-        return ResponseEntity.ok()
-                             .body(stationResponses);
+        return ResponseEntity.ok(stationResponses);
     }
 
     @DeleteMapping("/stations/{id}")
-    public ResponseEntity deleteStation(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteStation(@PathVariable Long id) {
         stationService.delete(id);
         return ResponseEntity.noContent()
                              .build();
