+++ src/test/java/wooteco/subway/station/StationAcceptanceTest.java
@@ -11,6 +11,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.jdbc.core.JdbcTemplate;
 import wooteco.subway.AcceptanceTest;
+import wooteco.subway.station.dto.StationResponse;
 
 import java.util.HashMap;
 import java.util.List;
